==> COAD_NormalCancer_Project/5_My_method_evaluation/2.Loading_TCGA_COAD_TPM_data_for_step1.R <==
#### 2.Loading_TCGA_COAD_TPM_data.R
#### read TCGA_COAD data ####
COAD_UCSC_Toil_tpm_dataset <- readRDS("/data8t_4/JH/MyJobs/Read_dataset/UCSC_Toil/COAD_UCSC_Toil_tpm_dataset_merge.syn2623706.rds")
COAD_RSEM_gene_tpm <- COAD_UCSC_Toil_tpm_dataset$COAD.RSEM.gene.tpm
gencode.v23.annotation <- COAD_UCSC_Toil_tpm_dataset$gencode.v23.annotation
COAD.pheno <- COAD_UCSC_Toil_tpm_dataset$COAD.pheno.merge.syn2623706
#head(gencode.v23.annotation)
## convert ensembleID to symbol 
geneMatch <- match(rownames(COAD_RSEM_gene_tpm),gencode.v23.annotation$V1)
geneSymbol <- as.character(gencode.v23.annotation[geneMatch,]$V2)
COAD_tpm_symbol <- COAD_RSEM_gene_tpm
rownames(COAD_tpm_symbol) <- geneSymbol
#head(COAD_tpm_symbol)
### Phenotype Update
### Seperate the Normal and Tumor samples
sampleID<- as.character(COAD.pheno$sample.x)
table(unlist(lapply(strsplit(sampleID,"-"), '[[', 4)))
sampleTypes <- unlist(lapply(strsplit(sampleID,"-"), '[[', 4))
sampleTypes[sampleTypes=="01"]<- "Tumor"
sampleTypes[sampleTypes=="11"]<- "Nomal"
COAD.pheno <- cbind(COAD.pheno , sampleTypes)

# Expression data transformation Log10(x+1)
COAD_tpm.Log10.expList <- list(COAD_tpm_symbol = log10(COAD_tpm_symbol+1))
COAD_tpm.Log10.expList.dataset <- list(COAD_tpm.Log10.expList = COAD_tpm.Log10.expList,
                                       COAD.pheno = COAD.pheno)
saveRDS(COAD_tpm.Log10.expList.dataset, file = "COAD_tpm.Log10.expList.dataset.rds")
